More precise mixin for entities
We now more precisely target laggier parts of entity processing rather than a shotgun approach. We now target the actual AI and entity collisions. We also now allow ticking on the client side for animations. This should resolve a lot of visual bugs with entities not being ticked and should also allow non-ticking entities to actually despawn and function. This causes a small drop in performance compared to before but is required to achieve vanilla parity.  We may later add a configuration to allow the previous implementation which is more aggressive, but it shouldn't be neccessary.