Better raid checks
We now check the raiders hasActiveRaid() function directly. We also now allow the villagers and iron golems in a village to tick in case of to give them a fair chance during a raid. Entities part of a raid are only exempt from ticking rules if the new config is enabled.