Jade Integration
There is now jade integration. This shows on entities which have been disabled by YouShallNotTick and on the totem of ticking when it has been disabled by the server configuration.
Spectators (by default) no longer count as players when calculating the nearby players for entities to tick.
Modified names of localisation strings.