Outline Renderer Backbone
Created the backbone for the new outline renderer to function. It's quite simply a vertex buffer which accepts vertices once and renders them in the world. It only generates the vertices/buffer once for performance. This may be expanded in future if needed.
This outline renderer is going to be used for rendering the radius of the totem of ticking. It is also flexible enough that it could be expanded to other things in future.
Also created an outline generator which is a class that can be dynamically added to and removed from the current outline renderer to add and remove vertices from the vertex buffer. It holds one method which is called when the buffers are updated in the renderer, making it quite modular.
Added embeddium and spark to the runtime environment on forge for testing.