Transformed Outlines
Fixed rendering order of outlines so transparency works as expected. On Fabric, we had to create mixin to do this. Outlines are now rendered right after blocks are rendered in the world.
The outline being rendered can now be transformed dynamically during render, allowing the creation of animations. This method is to be implemented by every OutlineGenerator. Added a temporary test pulse animation to the totem outline for now.
Refactored the outline renderer so that eventually the OutlineRenderer will be attached to the object which uses the outline, rather than being a single instance in the world. This has the drawback of making each OutlineRenderer have it's own draw call and only one generator attached, but is much easier to work with. This is probably going to changed again in the future, but for now this is acceptable.
Updated and reenabled Ponder so that it works on Fabric