Synced totem status maps
Added a new blockstate property to the totem of ticking, outlined. This defines whether the outlining rendering is currently active. The outlining rendering is toggled by sneak clicking the totem with an empty hand.
Added a new map which holds which totems have the outline rendering enabled.
Refactored the functions used for enabling/disabling totems to get rid of the update function and instead we just use add and remove directly.
Whenever a map is updated now, the new state is sent to all connected clients. Whenever a client joins the game the current serverside maps are sent to all clients just incase of desync.
Also fixed a bug that prevented more than one totem of ticking working upon chunk load.