Block entity renderers for outlines
Overhauled method for rendering outlines on the client side. The old way we were doing it was quite frankly awful, especially when BlockEntityRenderers (BER) exist. Stripped out all networking we added in last commit and replaced it with BERs. This drastically simplifies the rendering logic, as we no longer do this as a seperate rendering pass, we just initialise the renderers at startup.
We now construct the vertex data for outlines in global space not offset by the totem's position as that transform is automatically applied by the BER.
Fixed rendering order caused by new BER by enabling and disabling depth test before rendering outlines.
The outline now changes colour depending on if the totem is active or not. We are still using a test outlne.
Temporarily disabled ponder integration on Fabric as it fails to launch outside of a dev environment and also on a dedicated server.