Single totem outline generator instances
We now have a single instance of the totem outline generator instances for both the inactive and active state. The outline generator itself now holds the vertex buffer instead of the outline renderer. The reason for this is allowing the outline renderer to switch the generator (and hence the buffer) without having to regenerate the lot, this allows much more efficient switching of states in the outline.
As a result of this, the renderer classes are even simpler than before. Also removed the passing of the camera position and the block entity position as the block entity renderer automatically applies the transformations needed for this.