Totem outline animation
Added an animation for enabling the outline on a totem, as a result of this the way how the outline generator is set on the totem had to be reworked.
Adjusted colours for the outline to reflect the totem's model.
Fixed the outline not being completely centered on the totem's blockPos when it is scaled down.