Accurate outlines
Refactored the utils class to be a lot less cumbersome, removed a lot of the comments here as now it reads easy.
Fixed an issue where the scaling origin was not in the center of the totem, but was on the corner of the block.
Fixed an issue where the totem outline was offset incorrectly by 1 on the x axis.
We also now only use block positions within the distance checks so that the outline matches the actual ticking range, rather than having the outline not having the same degree of accuracy.