Fix outline culling on NeoForge
This is a little bit hacky, but it works. We've made a new version of the Ticking Totem Block Renderer specifically for NeoForge. This basically calls all the methods of the normal one, but allows the override function getRenderBoundingBox to be used which is added by NeoForge.
Also removed pack.mcmeta as NeoForge generates this at runtime now.
I believe this completes the port to 1.21.1.